+++ resources/data.csv
@@ -1,2 +1,3 @@
 aaaaaa,Ab-1sdf,2,gorda,3
 bbbbbb,Ab-1sdf,2,gorda,0
+ccccccc,Ab-1sdf,5,bs as,3
